+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.github.fertilewaif.chessboard"
+        applicationId "com.github.gmkornilov.chessboard"
         minSdkVersion 22
         targetSdkVersion 30
         versionCode 1
@@ -41,4 +41,6 @@ dependencies {
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+
+    implementation project(":chessboard")
 }
