+++ chessboard/src/androidTest/java/com/github/fertilewaif/chessboard/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.github.fertilewaif.chessboard
+package com.github.gmkornilov.chessboard
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.github.fertilewaif.chessboard.test", appContext.packageName)
+        assertEquals("com.github.gmkornilov.chessboard.test", appContext.packageName)
     }
 }
