+++ chessboard/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.github.fertilewaif.chessboard">
+    package="com.github.gmkornilov.chessboard">
 
 </manifest>
