--- chessboard/src/main/java/com/github/fertilewaif/chessboard/model/AnimationInfo.kt
@@ -1,5 +1,0 @@
-package com.github.fertilewaif.chessboard.model
-
-import androidx.annotation.DrawableRes
-
-data class AnimationInfo(@DrawableRes val drawable: Int, val from: CellInfo?, val to: CellInfo?)
