--- chessboard/src/main/java/com/github/fertilewaif/chessboard/model/FenParseException.kt
@@ -1,5 +1,0 @@
-package com.github.fertilewaif.chessboard.model
-
-import java.lang.Exception
-
-class FenParseException(message: String?) : Exception(message)
