--- chessboard/src/main/java/com/github/fertilewaif/chessboard/model/moves/RookTransitionMove.kt
@@ -1,9 +1,0 @@
-package com.github.fertilewaif.chessboard.model.moves
-
-import com.github.fertilewaif.chessboard.model.CellInfo
-import com.github.fertilewaif.chessboard.model.pieces.Piece
-import com.github.fertilewaif.chessboard.model.pieces.Rook
-
-class RookTransitionMove(piece: Rook, from: CellInfo, to: CellInfo) : TransitionMove(piece, from, to) {
-
-}
