+++ chessboard/src/main/java/com/github/gmkornilov/chessboard/model/AnimationInfo.kt
@@ -1,0 +1,5 @@
+package com.github.gmkornilov.chessboard.model
+
+import androidx.annotation.DrawableRes
+
+data class AnimationInfo(@DrawableRes val drawable: Int, val from: CellInfo?, val to: CellInfo?)
