+++ chessboard/src/main/java/com/github/gmkornilov/chessboard/model/FenParseException.kt
@@ -1,0 +1,5 @@
+package com.github.gmkornilov.chessboard.model
+
+import java.lang.Exception
+
+class FenParseException(message: String?) : Exception(message)
