+++ chessboard/src/main/java/com/github/gmkornilov/chessboard/model/moves/RookTransitionMove.kt
@@ -1,0 +1,9 @@
+package com.github.gmkornilov.chessboard.model.moves
+
+import com.github.gmkornilov.chessboard.model.CellInfo
+import com.github.gmkornilov.chessboard.model.pieces.Piece
+import com.github.gmkornilov.chessboard.model.pieces.Rook
+
+class RookTransitionMove(piece: Rook, from: CellInfo, to: CellInfo) : TransitionMove(piece, from, to) {
+
+}
